QWizard: Support fractional DPR on Windows
We were storing the DPR as int; change to qreal.

Task-number: QTBUG-114175
(cherry picked from commit c0e48ed645f68e714e982064540d710676a0738b)

+++ src/widgets/dialogs/qwizard_win_p.h
@@ -116,7 +116,7 @@ private:
 
     int titleBarOffset;  // Extra spacing above the text
     int iconSpacing;    // Space between button and icon
-    static int m_devicePixelRatio;
+    static qreal m_devicePixelRatio;
 };
